trace_id and session_id
