+++ app-server/src/agent_manager/agent_manager_grpc.rs
@@ -17,15 +17,13 @@ pub struct RunAgentRequest {
     pub request_api_key: ::core::option::Option<::prost::alloc::string::String>,
     #[prost(string, optional, tag = "4")]
     pub parent_span_context: ::core::option::Option<::prost::alloc::string::String>,
-    #[prost(string, optional, tag = "5")]
-    pub agent_state: ::core::option::Option<::prost::alloc::string::String>,
-    #[prost(enumeration = "ModelProvider", optional, tag = "6")]
+    #[prost(enumeration = "ModelProvider", optional, tag = "5")]
     pub model_provider: ::core::option::Option<i32>,
-    #[prost(string, optional, tag = "7")]
+    #[prost(string, optional, tag = "6")]
     pub model: ::core::option::Option<::prost::alloc::string::String>,
-    #[prost(bool, optional, tag = "8")]
+    #[prost(bool, optional, tag = "7")]
     pub enable_thinking: ::core::option::Option<bool>,
-    #[prost(message, repeated, tag = "9")]
+    #[prost(message, repeated, tag = "8")]
     pub cookies: ::prost::alloc::vec::Vec<Cookie>,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
@@ -45,15 +43,17 @@ pub struct StepChunkContent {
     pub action_result: ::core::option::Option<ActionResult>,
     #[prost(string, tag = "2")]
     pub summary: ::prost::alloc::string::String,
+    #[prost(string, tag = "3")]
+    pub trace_id: ::prost::alloc::string::String,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct AgentOutput {
-    #[prost(string, tag = "1")]
-    pub agent_state: ::prost::alloc::string::String,
-    #[prost(message, optional, tag = "2")]
+    #[prost(message, optional, tag = "1")]
     pub result: ::core::option::Option<ActionResult>,
-    #[prost(message, repeated, tag = "3")]
+    #[prost(message, repeated, tag = "2")]
     pub cookies: ::prost::alloc::vec::Vec<Cookie>,
+    #[prost(string, tag = "3")]
+    pub trace_id: ::prost::alloc::string::String,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct RunAgentResponseStreamChunk {
