+++ app-server/src/agent_manager/agent_manager_impl.rs
@@ -38,7 +38,6 @@ impl AgentManagerTrait for AgentManagerImpl {
         session_id: Option<Uuid>,
         request_api_key: Option<String>,
         parent_span_context: Option<String>,
-        agent_state: Option<String>,
         model_provider: Option<ModelProvider>,
         model: Option<String>,
         enable_thinking: bool,
@@ -51,7 +50,6 @@ impl AgentManagerTrait for AgentManagerImpl {
             session_id: session_id.map(|id| id.to_string()),
             request_api_key,
             parent_span_context,
-            agent_state,
             model_provider: model_provider.map(|p| p.to_i32()),
             model,
             enable_thinking: Some(enable_thinking),
@@ -69,7 +67,6 @@ impl AgentManagerTrait for AgentManagerImpl {
         session_id: Option<Uuid>,
         request_api_key: Option<String>,
         parent_span_context: Option<String>,
-        agent_state: Option<String>,
         model_provider: Option<ModelProvider>,
         model: Option<String>,
         enable_thinking: bool,
@@ -82,7 +79,6 @@ impl AgentManagerTrait for AgentManagerImpl {
             session_id: session_id.map(|id| id.to_string()),
             request_api_key,
             parent_span_context,
-            agent_state,
             model_provider: model_provider.map(|p| p.to_i32()),
             model,
             enable_thinking: Some(enable_thinking),
