+++ app-server/src/agent_manager/mock.rs
@@ -25,7 +25,6 @@ impl AgentManagerTrait for MockAgentManager {
         _session_id: Option<Uuid>,
         _request_api_key: Option<String>,
         _parent_span_context: Option<String>,
-        _agent_state: Option<String>,
         _model_provider: Option<ModelProvider>,
         _model: Option<String>,
         _enable_thinking: bool,
@@ -41,7 +40,6 @@ impl AgentManagerTrait for MockAgentManager {
         _session_id: Option<Uuid>,
         _request_api_key: Option<String>,
         _parent_span_context: Option<String>,
-        _agent_state: Option<String>,
         _model_provider: Option<ModelProvider>,
         _model: Option<String>,
         _enable_thinking: bool,
