+++ app-server/src/agent_manager/mod.rs
@@ -32,7 +32,6 @@ pub trait AgentManagerTrait {
         session_id: Option<Uuid>,
         request_api_key: Option<String>,
         parent_span_context: Option<String>,
-        agent_state: Option<String>,
         model_provider: Option<ModelProvider>,
         model: Option<String>,
         enable_thinking: bool,
@@ -45,7 +44,6 @@ pub trait AgentManagerTrait {
         session_id: Option<Uuid>,
         request_api_key: Option<String>,
         parent_span_context: Option<String>,
-        agent_state: Option<String>,
         model_provider: Option<ModelProvider>,
         model: Option<String>,
         enable_thinking: bool,
