+++ app-server/src/api/v1/agent.rs
@@ -69,7 +69,6 @@ pub async fn run_agent_manager(
                 None,
                 Some(request_api_key_vals.value),
                 request.parent_span_context,
-                None,
                 request.model_provider,
                 request.model,
                 request.enable_thinking,
@@ -92,7 +91,6 @@ pub async fn run_agent_manager(
                 None,
                 Some(request_api_key_vals.value),
                 request.parent_span_context,
-                None,
                 request.model_provider,
                 request.model,
                 request.enable_thinking,
